Move TypePack library headers into their own directory.
Includes associated build fixes.
