+++ src/osvr/CMakeLists.txt
@@ -5,6 +5,7 @@ include("${CMAKE_CURRENT_SOURCE_DIR}/GenerateForReportTypes.cmake")
 include("${CMAKE_CURRENT_SOURCE_DIR}/Macros.cmake")
 include("${CMAKE_CURRENT_SOURCE_DIR}/GenerateXMacro.cmake")
 
+add_subdirectory(TypePack)
 add_subdirectory(Util)
 add_subdirectory(Common)
