+++ tests/cplusplus/Util/CMakeLists.txt
@@ -1,4 +1,4 @@
-foreach(testname TreeNode TypePack ContainerWrapper UniqueContainer Projection)
+foreach(testname TreeNode ContainerWrapper UniqueContainer Projection)
     add_executable(${testname} ${testname}.cpp)
     target_link_libraries(${testname} osvrUtilCpp)
     osvr_setup_gtest(${testname})
